+++ osquery/tables/CMakeLists.txt
@@ -14,6 +14,7 @@ if(APPLE)
     events/darwin/file_changes.cpp
     events/darwin/hardware_events.cpp
     events/darwin/yara.cpp
+    events/yara_utils.cpp
     networking/darwin/routes.cpp
     system/darwin/acpi_tables.cpp
     system/darwin/apps.cpp
@@ -61,6 +62,7 @@ else()
     events/linux/passwd_changes.cpp
     events/linux/file_changes.cpp
     events/linux/yara.cpp
+    events/yara_utils.cpp
     networking/linux/arp_cache.cpp
     networking/linux/process_open_sockets.cpp
     networking/linux/routes.cpp
